Update IoT_Button.ino
The led was turned on all the time and turned off when the button was pressed. It should have been the other way around I think.
 I also have another question apart from this proposal:
When I write this program to my nodemcu esp8266 development board, and when I run it the serial monitor just reads jibberish. The baudrate is set right but still... This means that I can't see what is not working.